requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
reportlab==4.0.9
openpyxl==3.1.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-cleanup==8.0.0
dj-database-url==2.1.0
